% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumtab.R
\name{sumtab}
\alias{sumtab}
\title{Create Summary Table}
\usage{
sumtab(
  data,
  by = NA,
  forced_nonparam = FALSE,
  analysis = TRUE,
  complete_rows = TRUE,
  same_row = TRUE,
  multivariate = FALSE,
  debug = FALSE
)
}
\arguments{
\item{data}{The dataset to be analyzed and summarized.}

\item{by}{A variable to group the data by; if set to the default, \code{NA}, no analysis will be performed.}

\item{forced_nonparam}{Logical. If \code{TRUE}, forces non-parametric analysis; otherwise, the choice between parametric and non-parametric methods is determined by conventional criteria. Default is \code{FALSE}.}

\item{analysis}{Logical. If \code{TRUE}, performs statistical analysis; if \code{FALSE}, generates descriptive statistics only. Default is \code{TRUE}.}

\item{complete_rows}{Logical. If \code{TRUE}, and if the feature is a factor with 2 levels, both levels will be shown in the table. Default is \code{TRUE}.}

\item{same_row}{Logical. If \code{TRUE}, displays the feature name in the same row as its summary data. Default is \code{TRUE}.}

\item{multivariate}{Logical. If \code{TRUE}, performs multivariate analysis. Default is \code{FALSE}.}

\item{debug}{Logical. If \code{TRUE}, enables debug mode and disables simplification. Default is \code{FALSE}.}
}
\value{
A string matrix with the requested analysis and format.
}
\description{
This function generates summary table given various customizable options.
}
\examples{
# Basic usage:
sumtab(data = my_data)
my_data \%>\% sumtab()

}
