% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumtab.R
\name{sumtab}
\alias{sumtab}
\title{Create Summary Table}
\usage{
sumtab(
  data,
  by = NA,
  reporting_type = "auto",
  analysis = TRUE,
  complete_rows = TRUE,
  same_row = TRUE,
  multivariate = FALSE,
  debug = FALSE,
  risk_measure = "OR",
  fmt_p_default = handle_ps,
  fmt_p_after = NULL,
  fmt_cd = default_fmt_cd
)
}
\arguments{
\item{data}{The dataset to be analyzed and summarized.}

\item{by}{A variable to group the data by; if set to the default, \code{NA}, no analysis will be performed.}

\item{reporting_type}{A string indicating the type of test to perform. Options are:
\describe{
\item{"parametric"}{Forces the function to perform a parametric test (e.g., t-test) and report the mean ± SD.}
\item{"non_parametric"}{Forces the function to perform a non-parametric test (e.g., Wilcoxon test) and report the median ± IQR.}
\item{"auto"}{Automatically chooses the test based on the data's characteristics
(e.g., a normality test to decide between parametric and non-parametric). If parametric, mean ± SD is reported;
if non-parametric, median ± IQR is reported.}
}}

\item{analysis}{Logical. If \code{TRUE}, performs statistical analysis; if \code{FALSE}, generates descriptive statistics only. Default is \code{TRUE}.}

\item{complete_rows}{Logical. If \code{TRUE}, and if the feature is a factor with 2 levels, both levels will be shown in the table. Default is \code{TRUE}.}

\item{same_row}{Logical. If \code{TRUE}, displays the feature name in the same row as its summary data. Default is \code{TRUE}.}

\item{multivariate}{Logical. If \code{TRUE}, performs multivariate analysis. Default is \code{FALSE}.}

\item{debug}{Logical. If \code{TRUE}, enables debug mode and disables simplification. Default is \code{FALSE}.}

\item{risk_measure}{A string indicating the type of association measure to calculate. Options are:
\describe{
\item{\code{"OR"}}{Calculates the Odds Ratio (OR) and its 95-percent confidence interval based on a 2x2 table.}
\item{\code{"RR"}}{Calculates the Relative Risk (RR) and its 95-percent confidence interval based on a 2x2 table.}
}}

\item{fmt_p_default}{A function applied to to each raw numeric p-value before
any user-specified post-processing. The default is the internal formatter \code{handle_ps()}.}

\item{fmt_p_after}{Optional function applied to each already-formatted
p-value. Must take a single character p-value and return a single character
value. If \code{NULL}, no post-formatting is performed.}

\item{fmt_cd}{A function used to format a pair of summary
statistics representing central tendency and dispersion (e.g. mean–SD or
median–IQR). The function must accept exactly two arguments
\code{(central, dispersion)},where each argument is already rounded
and formatted as a string, and return a single character scalar. The default is the internal formatter.}
}
\value{
A string matrix with the requested analysis and format.
}
\description{
This function generates summary table given various customizable options.
}
\examples{
# Basic usage:
sumtab(data = my_data)
my_data \%>\% sumtab()

}
